Compute the ensemble mean and variance

The ensemble mean and variance are computed and added as columns to tables in a harp_fcst object.

ens_mean_and_var(.fcst, mean_name = "ens_mean", var_name = "ens_var", ...)

Arguments

.fcst
A harp_fcst object, or a data frame containing columns with "_mbr" as part of the column names.

Value

A harp_fcst object with columns ens_mean and ens_var added to the forecast tables.