Show built in file templates

Show built in file templates

Usage

show_file_templates(template_row = NULL)

Arguments

template_row

Some of the templates are too wide to show on the screen, so to show a template in full, pass the row number of the template you want to show in full in template_row

Examples

show_file_templates()
#> # A tibble: 34 × 2
#>    template_name          template                                              
#>    <chr>                  <chr>                                                 
#>  1 arome_arctic_extracted /lustre/storeB/immutable/archive/projects/metproducti…
#>  2 arome_arctic_full      /lustre/storeB/immutable/archive/projects/metproducti…
#>  3 arome_arctic_sfx       /lustre/storeB/immutable/archive/projects/metproducti…
#>  4 fctable                {file_path}/{fcst_model}/{YYYY}/{MM}/FCTABLE_{paramet…
#>  5 fctable_det            {file_path}/{det_model}/{YYYY}/{MM}/FCTABLE_{paramete…
#>  6 fctable_eps            {file_path}/{eps_model}/{YYYY}/{MM}/FCTABLE_{paramete…
#>  7 fctable_eps_all_cycles {file_path}/{eps_model}/{YYYY}/{MM}/FCTABLE_{paramete…
#>  8 fctable_eps_all_leads  {file_path}/{eps_model}/{YYYY}/{MM}/FCTABLE_{paramete…
#>  9 glameps_grib           {file_path}/{eps_model}/{sub_model}/{YYYY}/{MM}/{DD}/…
#> 10 harmoneps_grib         {file_path}/{YYYY}/{MM}/{DD}/{HH}/mbr{MBR3}/fc{YYYY}{…
#> 11 harmoneps_grib_fp      {file_path}/{YYYY}/{MM}/{DD}/{HH}/mbr{MBR3}/fc{YYYY}{…
#> 12 harmoneps_grib_sfx     {file_path}/{YYYY}/{MM}/{DD}/{HH}/mbr{MBR3}/fc{YYYY}{…
#> 13 harmonie_grib          {file_path}/{YYYY}/{MM}/{DD}/{HH}/fc{YYYY}{MM}{DD}{HH…
#> 14 harmonie_grib_fp       {file_path}/{YYYY}/{MM}/{DD}/{HH}/fc{YYYY}{MM}{DD}{HH…
#> 15 harmonie_grib_sfx      {file_path}/{YYYY}/{MM}/{DD}/{HH}/fc{YYYY}{MM}{DD}{HH…
#> 16 meps_cntrl_extracted   /lustre/storeB/immutable/archive/projects/metproducti…
#> 17 meps_cntrl_sfx         /lustre/storeB/immutable/archive/projects/metproducti…
#> 18 meps_det               /lustre/storeB/immutable/archive/projects/metproducti…
#> 19 meps_extracted         /lustre/storeB/immutable/archive/projects/metproducti…
#> 20 meps_full              /lustre/storeB/immutable/archive/projects/metproducti…
#> 21 meps_lagged_6h_subset  /lustre/storeB/immutable/archive/projects/metproducti…
#> 22 meps_sfx               /lustre/storeB/immutable/archive/projects/metproducti…
#> 23 meps_subset            /lustre/storeB/immutable/archive/projects/metproducti…
#> 24 obsoul                 {file_path}/obsoul_1_xxxxxy_{country}_{YYYY}{MM}{DD}{…
#> 25 obstable               {file_path}/OBSTABLE_{YYYY}.sqlite                    
#> 26 vfld                   {file_path}/{fcst_model}/vfld{fcst_model}{YYYY}{MM}{D…
#> 27 vfld_det               {file_path}/{det_model}/vfld{det_model}{YYYY}{MM}{DD}…
#> 28 vfld_det_noexp         {file_path}/{det_model}/vfld{YYYY}{MM}{DD}{HH}{LDT2}  
#> 29 vfld_eps               {file_path}/{sub_model}/vfld{sub_model}mbr{MBR3}{YYYY…
#> 30 vfld_eps_noexp         {file_path}/{sub_model}/vfldmbr{MBR3}{YYYY}{MM}{DD}{H…
#> 31 vfld_multimodel        {file_path}/{sub_model}/vfld{sub_model}mbr{MBR3}{YYYY…
#> 32 vfld_multimodel_noexp  {file_path}/{sub_model}/vfldmbr{MBR3}{YYYY}{MM}{DD}{H…
#> 33 vfld_noexp             {file_path}/{fcst_model}/vfldmbr{MBR3}{YYYY}{MM}{DD}{…
#> 34 vobs                   {file_path}/vobs{YYYY}{MM}{DD}{HH}                    
show_file_templates(1)
#> 
#> template_name:
#>  arome_arctic_extracted 
#>  
#> template:
#>  /lustre/storeB/immutable/archive/projects/metproduction/DNMI_AROME_ARCTIC/{YYYY}/{MM}/{DD}/arome_arctic_extracted_2_5km_{YYYY}{MM}{DD}T{HH}Z.nc 
show_file_templates(18)
#> 
#> template_name:
#>  meps_det 
#>  
#> template:
#>  /lustre/storeB/immutable/archive/projects/metproduction/MEPS/{YYYY}/{MM}/{DD}/meps_det_2_5km_{YYYY}{MM}{DD}T{HH}Z.nc