Read gridded data in various formats

read_grid is a generic function to read gridded data from a file. A transformation can optionally be applied to those data.

read_grid(
  file_name,
  parameter,
  file_format = NULL,
  file_format_opts = list(),
  vertical_coordinate = c("pressure", "model", "height", NA),
  lead_time = NULL,
  members = NULL,
  transformation = c("none", "interpolate", "regrid", "xsection"),
  transformation_opts = list(),
  show_progress = FALSE,
  data_frame = FALSE,
  readable_times = TRUE,
  spread_members = FALSE,
  ...
)

read_grid_interpolate(file_name, parameter, file_format = NULL, ...)

Arguments

file_name
A character string with the (full) file name.

parameter
The parameter to be read.

file_format
The file format. Possible values include grib, netcdf, FA, hdf5... Whatever the value is, it is suposed to correspond to a function "read_XXX" that can deal with the format. If not specified, the format can often be guessed correctly from file extension or the first few bytes.

file_format_opts
Options for the file format as generated by one of the options functions, e.g. netcdf_opts.

vertical_coordinate
The vertical coordinate for upper air data. May be "pressure" for pressure levels, "model" for model levels or "height" for height levels.

lead_time
The lead times to read from a forecast file. If set to NULL, all lead times will be read from the file.

members
For ensemble data, the ensemble members to read from the file. If set to NULL all members will be read.

transformation
The transformation to apply to the gridded data. Can be "none", "interpolate", "regrid" or "xsection".

transformation_opts
Options for the transformation as generated by interpolate_opts, regrid_opts or xsection_opts.

show_progress
Show progress when reading large amounts of data.

data_frame
Logical. For compatibility with current version of harpSpatial, make sure only a geofield is returned rather than a data frame when set to FALSE.

readable_times
When data_frame = TRUE, whether to convert times in unix format to a data-time format. The default is TRUE.

...
All arguments passed to the specified reader function.

Value

A geofield or (possibly) a plain matrix.

Examples

if (requireNamespace("Rgrib2", quietly = TRUE) & requireNamespace("harpData", quietly = TRUE)) {
  read_grid(
    system.file("grib/HARMUK20171015T12Z+003.grib", package = "harpData"),
    parameter = "T2m"
  )
  read_grid(
    system.file("grib/HARMUK20171015T12Z+003.grib", package = "harpData"),
    parameter = "RH2m"
  )
  read_grid(
    system.file("grib/HARMUK20171015T12Z+003.grib", package = "harpData"),
    parameter = "tcc"
  )
}